+++ server/streams/consumer.go
@@ -7,16 +7,22 @@ import (
 	"github.com/go-redis/redis"
 )
 
-func (r *RedisClient) Read(stream, group string, pending bool) (result []redis.XStream, err error) {
+func (r *RedisClient) Read(stream, group string, count int64, pending bool) (result []redis.XStream, err error) {
 	IDs := ">"
 	if pending {
 		IDs = "0"
 	}
-	result, err = r.DB.XReadGroup(&redis.XReadGroupArgs{
+
+	args := redis.XReadGroupArgs{
 		Streams: []string{stream, IDs},
 		Group:   group,
 		Block:   1 * time.Second,
-	}).Result()
+	}
 
+	if count != 0 {
+		args.Count = count
+	}
+
+	result, err = r.DB.XReadGroup(&args).Result()
 	return
 }
